backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42
pydantic-settings==2.5.2
httpx==0.28.1
python-dotenv==1.0.1
pytest==9.0.2
pytest-asyncio==1.3.0
pydantic==2.11.7
ollama==0.6.1